gui: satisfy the qt-includes and locale-dependence lint gates
The -devsyntheticcoins harness tripped two hard-fail gates in CI lint:

- lint-qt-includes (the Phase 1f gate): src/qt/syntheticcoinsource.h
  held GRC::WalletCoinEventQueue and GRC::WalletCoinStore by value, so it
  had to include wallet/wallet_event_queue.h and wallet/walletcoinstore.h
  from a src/qt header outside the composition root. Move the harness
  node-side to wallet/synthetic_coin_source.cpp behind a new
  interfaces::MakeSyntheticCoinSource(total, groups) factory -- the same
  shape as MakeWalletCoinSource, with the implementation class private to
  the .cpp. The GUI now reaches it through interfaces/ like every other
  source, and bitcoin.cpp drops the qt-header include.

- lint-locale-dependence: std::atoi parsing the argument. Use ParseInt32
  and throw on a malformed argument rather than silently harnessing the
  wrong scale.

Also fix one codespell hit in walletcoinstore.h (direcory -> directory).
The remaining codespell output in that job is pre-existing on development
and warn-only (lint-spelling.sh always exits 0).
