+++ src/CMakeLists.txt
@@ -263,6 +263,7 @@ add_library(gridcoin_util STATIC
     wallet/interfaces.cpp
     wallet/rpcdump.cpp
     wallet/rpcwallet.cpp
+    wallet/synthetic_coin_source.cpp
     wallet/transactionrecord.cpp
     wallet/txdetail.cpp
     wallet/txfilter.cpp
